* main: (69 commits)
  The DeviceFeatureLimits::BufferAlignment is not always equals 16 byte on OpenGL backend. (#113)
  Suppress swiftshader's TSAN false-positives
  Remove OpenGL dependency from Vulkan tests
  Disable debug names on Android due to emulator crash
  igl | vulkan - scope guard for locked hw buffer
  igl | vulkan - external hw buffer support #2
  igl | vulkan - external hw buffer support #1
  igl - native hw buffer desc helper
  igl | vulkan - native hw buffer
  igl | vulkan - hw native buffer external memory support
  igl - native hw buffer interface
  add multiviewmultisample devicefeature check
  Assert even if we don't terminate to allow debug break on validation error
  Fixed shell tests compilation
  igl | shell | Demonstrate both YUV formats `NV12` and `420p`.
  igl | shell | Enqueue key events on Windows
  igl | Add helper function `SamplerStateDesc::newYUV()`
  Bump the `apk` modes to NDK 26, API 31 and update Hyperspace
  igl | Add YUV 420p test data file
  igl | Add new texture format `YUV_420p`
  ...

# Conflicts:
#	src/igl/opengl/egl/PlatformDevice.cpp

+++ shell/shared/testShell/TestShell.h
@@ -15,8 +15,8 @@
 namespace igl::shell {
 
 struct ScreenSize {
-  uint width;
-  uint height;
+  size_t width;
+  size_t height;
 };
 
 class TestShellBase {
